Developer to master with notification drop-down phone bar for a current song that is played.

+++ Music app/app/src/main/java/com/example/musicapp/Musics/MusicAdapter.java
@@ -8,11 +8,9 @@ import android.media.MediaMetadataRetriever;
 import android.net.Uri;
 import android.provider.MediaStore;
 import android.view.LayoutInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.PopupMenu;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
@@ -20,6 +18,7 @@ import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
+import com.example.musicapp.Activities.MainActivity;
 import com.example.musicapp.Activities.PlayerActivity;
 import com.example.musicapp.R;
 import com.google.android.material.snackbar.Snackbar;
@@ -32,21 +31,13 @@ import java.util.ArrayList;
 
 public class MusicAdapter extends RecyclerView.Adapter<MusicAdapter.MyViewHolder>
 {
-    private final Context mContext;
     public static ArrayList<MusicFile> mFiles;
+    private final Context mContext;
 
     public MusicAdapter(Context mContext, ArrayList<MusicFile> mFiles)
     {
         this.mContext = mContext;
-        this.mFiles = mFiles;
-    }
-
-    @NonNull
-    @Override
-    public MyViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType)
-    {
-        View view = LayoutInflater.from(mContext).inflate(R.layout.music_items, parent, false);
-        return new MyViewHolder(view);
+        MusicAdapter.mFiles = mFiles;
     }
 
     public static boolean isPathValid(String path)
@@ -61,6 +52,23 @@ public class MusicAdapter extends RecyclerView.Adapter<MusicAdapter.MyViewHolder
         }
 
         return true;
+    }
+
+    public static byte[] getAlbumArt(String uri)
+    {
+        MediaMetadataRetriever retriever = new MediaMetadataRetriever();
+        retriever.setDataSource(uri);
+        byte[] art = retriever.getEmbeddedPicture();
+        retriever.release();
+        return art;
+    }
+
+    @NonNull
+    @Override
+    public MyViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType)
+    {
+        View view = LayoutInflater.from(mContext).inflate(R.layout.music_items, parent, false);
+        return new MyViewHolder(view);
     }
 
     @Override
@@ -108,49 +116,44 @@ public class MusicAdapter extends RecyclerView.Adapter<MusicAdapter.MyViewHolder
         holder.menuMore.setOnClickListener(new View.OnClickListener()
         {
             @Override
-            public void onClick(final View v)
+            public void onClick(final View view)
             {
-                PopupMenu popupMenu = new PopupMenu(mContext, v);
-                popupMenu.getMenuInflater().inflate(R.menu.popup, popupMenu.getMenu());
-                popupMenu.show();
-                popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener()
+                if (PlayerActivity.curr_song != null && PlayerActivity.curr_song == mFiles.get(position))
                 {
-                    @Override
-                    public boolean onMenuItemClick(MenuItem item)
-                    {
-                        if (item.getItemId() == R.id.delete)
-                            deleteFile(position, v);
+                    Snackbar.make(view, "Can't be deleted while playing!", Snackbar.LENGTH_LONG).show();
+                    return;
+                }
 
-                        return true;
-                    }
-                });
+                Uri contentUri = ContentUris.withAppendedId(MediaStore.Audio.Media.EXTERNAL_CONTENT_URI,
+                        Long.parseLong(mFiles.get(position).getId()));
+                File file = new File(mFiles.get(position).getPath());
+                boolean deleted = file.delete();
+                if (deleted)
+                {
+                    mContext.getContentResolver().delete(contentUri, null, null);
+                    mFiles.remove(position);
+                    notifyItemRemoved(position);
+                    notifyItemRangeChanged(position, mFiles.size());
+                    MainActivity.musicFiles.remove(position);
+                    Snackbar.make(view, "File deleted", Snackbar.LENGTH_LONG).show();
+                }
+                else
+                    Snackbar.make(view, "Can't be deleted!", Snackbar.LENGTH_LONG).show();
             }
         });
-    }
-
-    public void deleteFile(int position, View view)
-    {
-        Uri contentUri = ContentUris.withAppendedId(MediaStore.Audio.Media.EXTERNAL_CONTENT_URI,
-                Long.parseLong(mFiles.get(position).getId()));
-        File file = new File(mFiles.get(position).getPath());
-        boolean deleted = file.delete(); // delete file
-        if (deleted)
-        {
-            mContext.getContentResolver().delete(contentUri, null, null);
-            mFiles.remove(position);
-            notifyItemRemoved(position);
-            notifyItemRangeChanged(position, mFiles.size());
-            Snackbar.make(view, "File deleted", Snackbar.LENGTH_LONG).show();
-        }
-        // else is when the file is in the sd card
-        else
-            Snackbar.make(view, "Can't be deleted!", Snackbar.LENGTH_LONG).show();
     }
 
     @Override
     public int getItemCount()
     {
         return mFiles.size();
+    }
+
+    public void updateList(ArrayList<MusicFile> musicFileArrayList)
+    {
+        mFiles = new ArrayList<>();
+        mFiles.addAll(musicFileArrayList);
+        notifyDataSetChanged();
     }
 
     public static class MyViewHolder extends RecyclerView.ViewHolder
@@ -164,23 +167,7 @@ public class MusicAdapter extends RecyclerView.Adapter<MusicAdapter.MyViewHolder
             file_name = itemView.findViewById(R.id.music_file_name);
             album_name = itemView.findViewById(R.id.album_file_name);
             album_art = itemView.findViewById(R.id.music_img);
-            menuMore =  itemView.findViewById(R.id.menuMore);
+            menuMore = itemView.findViewById(R.id.menuMore);
         }
-    }
-
-    public static byte[] getAlbumArt(String uri)
-    {
-        MediaMetadataRetriever retriever = new MediaMetadataRetriever();
-        retriever.setDataSource(uri);
-        byte[] art = retriever.getEmbeddedPicture();
-        retriever.release();
-        return art;
-    }
-
-    public void updateList(ArrayList<MusicFile> musicFileArrayList)
-    {
-        mFiles = new ArrayList<>();
-        mFiles.addAll(musicFileArrayList);
-        notifyDataSetChanged();
     }
 }
